[android] Inlined methods, added suppress annotations https://github.com/mapsme/omim/pull/10273#discussion_r252117813 https://github.com/mapsme/omim/pull/10273#discussion_r252117511 https://github.com/mapsme/omim/pull/10273#discussion_r252117215
