+++ android/src/com/mapswithme/maps/base/BaseAuthFragment.java
@@ -1,6 +1,5 @@
 package com.mapswithme.maps.base;
 
-import android.content.Context;
 import android.content.Intent;
 import android.support.annotation.CallSuper;
 import android.support.annotation.NonNull;
@@ -8,8 +7,14 @@ import android.support.annotation.Nullable;
 
 import com.mapswithme.maps.auth.Authorizer;
 import com.mapswithme.maps.auth.TargetFragmentCallback;
+import com.mapswithme.maps.base.BaseMwmToolbarFragment;
 
-public abstract class BaseAuthFragment extends BaseAsyncOperationFragment
+/**
+ * A base toolbar fragment which is responsible for the <b>authorization flow</b>,
+ * starting from the getting an auth token from a social network and passing it to the core
+ * to get user authorized for the MapsMe server (Passport).
+ */
+public abstract class BaseAuthFragment extends BaseMwmToolbarFragment
     implements Authorizer.Callback, TargetFragmentCallback
 {
   @NonNull
@@ -22,17 +27,17 @@ public abstract class BaseAuthFragment extends BaseAsyncOperationFragment
 
   @Override
   @CallSuper
-  public void onAttach(Context context)
+  public void onStart()
   {
-    super.onAttach(context);
+    super.onStart();
     mAuthorizer.attach(this);
   }
 
   @Override
   @CallSuper
-  public void onDestroyView()
+  public void onStop()
   {
-    super.onDestroyView();
+    super.onStop();
     mAuthorizer.detach();
   }
