+++ android/src/com/mapswithme/maps/ugc/routes/SendLinkPlaceholderFragment.java
@@ -9,7 +9,7 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.mapswithme.maps.R;
-import com.mapswithme.maps.base.BaseAuthFragment;
+import com.mapswithme.maps.base.BaseToolbarAuthFragment;
 import com.mapswithme.maps.bookmarks.data.BookmarkCategory;
 import com.mapswithme.maps.bookmarks.data.BookmarkManager;
 import com.mapswithme.maps.bookmarks.data.CatalogCustomProperty;
@@ -21,7 +21,7 @@ import com.mapswithme.util.statistics.Statistics;
 import java.util.List;
 import java.util.Objects;
 
-public class SendLinkPlaceholderFragment extends BaseAuthFragment implements BookmarkManager.BookmarksCatalogListener
+public class SendLinkPlaceholderFragment extends BaseToolbarAuthFragment implements BookmarkManager.BookmarksCatalogListener
 {
   public static final String EXTRA_CATEGORY = "bookmarks_category";
   private static final String BODY_STRINGS_SEPARATOR = "\n\n";
@@ -43,6 +43,7 @@ public class SendLinkPlaceholderFragment extends BaseAuthFragment implements Boo
     mCategory = Objects.requireNonNull(args.getParcelable(EXTRA_CATEGORY));
   }
 
+  @SuppressWarnings("NullableProblems")
   @Nullable
   @Override
   public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container,
@@ -95,16 +96,11 @@ public class SendLinkPlaceholderFragment extends BaseAuthFragment implements Boo
     if (uploadResult == BookmarkManager.UploadResult.UPLOAD_RESULT_SUCCESS)
       onUploadSucceeded();
     else if (uploadResult == BookmarkManager.UploadResult.UPLOAD_RESULT_AUTH_ERROR)
-      onAuthError();
+      authorize();
     else
       onUploadFailed();
 
     hideProgress();
-  }
-
-  private void onAuthError()
-  {
-    authorize();
   }
 
   private void onUploadFailed()
